Move all CLI related code into dedicate module
